+++ src/Plugins/ef-querylens-vscode/src/test/suite/sqlReady.test.ts
@@ -183,5 +183,6 @@ function createSettings(): QueryLensSettings {
         hoverWaitWhenWarmMs: 0,
         hoverProgressNotify: false,
         notifyWhenSqlReady: true,
+        dotnetPath: '',
     };
 }